Fix hue ring rendering outside of it's bounds
The hue ring was rendering outside of it's bounds because drawCircle
draws the stroke width half before the radius and the other half after.

This commit fixes that by subtracting half the stroke width from the
radius.

HueRingPicker was adjusted to retain it's previous UI and to allow for
larger stroke widths without clipping