Allow to only recognize inputs inside the hue ring
Adds a new option to toggle between the always win pan gesture recognizer with a
new recognizer that checks if the pointer event is inside the ring.

This allows direct users of the `ColorPickerHueRing` to be able to let
events that are in the bounds of the hue ring but not in the ring to
reach parents of the widget.